+++ euchrecli/__init__.py
@@ -1,2 +1,2 @@
 # release version
-version = '0.0.1'
+version = '0.1.0'