+++ euchrecli/util/deck_util.py
@@ -1,10 +1,13 @@
 
 from random import shuffle
-from euchrecli.card_util import Suit, Face, Card
-from euchrecli.player_util import Player
+from euchrecli.util.card_util import Suit, Face, Card
+from euchrecli.util.player_util import Player
 
 
 def create_deck() -> [Card]:
+    """Return a shuffled standard euchre deck.
+    Creates a deck containing 9, 10, J, Q, K, A of each suit.
+    """
     suits = [
         Suit('Spade', 'Black'),
         Suit('Club', 'Black'),
@@ -29,7 +32,10 @@ def create_deck() -> [Card]:
     shuffle(deck)
     return deck
 
+
 def deal_hand(players: [Player], deck: [Card]) -> None:
+    """Deal five cards from the deck to each player."""
+    # TODO: implement standard euchre dealing algorithm
     index = 0
     for player in players:
         for _ in range(5):