+++ CHANGELOG.md
@@ -4,3 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.1.0] - 2020-07-06
+
+### Added
+
+- `euchre play` - Four computer players can successfully play a full game of
+    euchre together, making random decisions (within the rules)
+- `euchre --version` - Check current installed version
+- `euchre --help` - Print out cli usage